feat(past_usage): Refact current usage
