+++ app/controllers/api/v1/customers_controller.rb
@@ -17,29 +17,6 @@ module Api
         end
       end
 
-      def current_usage
-        service = ::Invoices::CustomerUsageService
-          .new(
-            nil,
-            customer_id: params[:customer_external_id],
-            subscription_id: params[:external_subscription_id],
-            organization_id: current_organization.id,
-          )
-        result = service.usage
-
-        if result.success?
-          render(
-            json: ::V1::CustomerUsageSerializer.new(
-              result.usage,
-              root_name: 'customer_usage',
-              includes: %i[charges_usage],
-            ),
-          )
-        else
-          render_error_response(result)
-        end
-      end
-
       def portal_url
         customer = current_organization.customers.find_by(external_id: params[:customer_external_id])
